// src/app/modules/openapi/model/user-sign-up-dto.ts
/**
 * E-Commerce Application
 *
 * Contact: sad@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserSignUpDTO { 
    userName: string;
    email: string;
    password: string;
    BD?: string;
    job?: string;
    creditNo?: string;
    creditLimit?: number;
    phone?: string;
    bd?: string;
}